1.完善用户的资料编辑 2.升级到 act v1.8.7

+++ src/main/java/com/artlongs/sys/dao/SysPermissionDao.java
@@ -42,7 +42,7 @@ public class SysPermissionDao extends SysPermission.Dao<SysPermission> {
         return permissionList;
     }
 
-    public SysPermission getPermissionOf(Integer funcid, Integer roleid) {
+    public SysPermission getPermissionOf(Long funcid, Long roleid) {
 //        String sql = " select * from sys_permission where func_id=? and role_id = ?";
 //        SysPermission sysPermission = getObj(sql, funcId, roleId);
         String sql = new Qe(SysPermission.class).where(new Qe().eq(funcId,funcid).and(new Qe().eq(roleId,roleid))).build();
